+++ app/settings/contentCard-settings-otherApis.php
@@ -1,6 +1,6 @@
 <?php
 // Card
-$cardHeaderContent = \xan\iconFA( 'fas fa-external-link-square-alt' ) . STR_NBSP . 'Other APIs';
+$cardHeaderContent = \xan\fontIcon( 'fas fa-external-link-square-alt' ) . STR_NBSP . 'Other APIs';
 $card = new \xan\eleCard( CARD_WIDTH, '', false );
 
 // Tags Special