+++ app/zzMetaModules/mmUsersLogin.php
@@ -21,8 +21,8 @@ class moduleMetaUsersLogin extends \xan\moduleMeta {
 		$this->NamePlural = 'Login';
 		$this->NameSingular = 'Login';
 		
-		$this->FontAwesome = '<i class=\'fas fa-door-open\'></i>';
-		$this->FontAwesomeList = FA_LIST;
+		$this->FontAwesome = \xan\fontIcon( 'fas fa-door-open' );
+		$this->FontAwesomeList = FI_LIST;
 		
 		$this->URLRelative = '/' . 'login/';
 		$this->URLFull = URL_BASE . 'logout/';