+++ templates/page-resp.php
@@ -95,7 +95,9 @@ ob_start();
 			echo "\r\r";
 		}
 		?>
-
+        <!-- Bootstrap Icons     https://icons.getbootstrap.com -->
+        <link href="/include/bootstrapIcons/1.5.0/font/bootstrap-icons.css" rel="stylesheet" type="text/css">
+        
         <!-- FontAwesome     https://fontawesome.com/icons -->
         <link href="/include/fontawesome/5.15.2_pro/css/all.min.css" rel="stylesheet" type="text/css">
 
@@ -348,7 +350,7 @@ ob_start();
 				<?php if ( $resp->navInclude ) : ?>
 
                     <!-- Menu Hamburger -->
-                    <button class="navbar-toggler nav-hamburger" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation"><?= FA_MENU . STR_NBSP ?>Menu</button>
+                    <button class="navbar-toggler nav-hamburger" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation"><?= FI_MENU . STR_NBSP ?>Menu</button>
 
                     <!-- Menu Expanded -->
                     <div class="collapse navbar-collapse" id="navbarSupportedContent">
@@ -372,7 +374,7 @@ ob_start();
 										echo \xan\navDivider();
 									}
 									if ( \xan\isNotEmpty( $_SESSION[ SESS_USER ][ UUIDUSERS ] ) ) {
-										echo \xan\navItemDropdownCustom( FA_PASSWORD . STR_NBSP . 'Change My Password', /** @lang JavaScript */ "$('#UsersPasswordChange_Modal').modal('show');" );
+										echo \xan\navItemDropdownCustom( FI_PASSWORD . STR_NBSP . 'Change My Password', /** @lang JavaScript */ "$('#UsersPasswordChange_Modal').modal('show');" );
 										echo \xan\navDivider();
 										require_once( PATH_ROOT_APP . 'settings-users/contentModal-users-passwordChange.php' );
 									}
@@ -429,7 +431,7 @@ ob_start();
     <script>
         function xanDo( params ) {
             let iconID = Math.floor( ( Math.random() * 999999999 ) + 1 );
-            xanMessageDisplay( "#xanMessage", "<?= FA_PSOS_CLOUD ?>", iconID, '', false, false );
+            xanMessageDisplay( "#xanMessage", "<?= FI_PSOS_CLOUD ?>", iconID, '', false, false );
             // alert( "xanDo params = " + JSON.stringify( params ) );
             let timeBegin = new Date();
             // Window in Current or New Window
@@ -540,7 +542,7 @@ ob_start();
                     xanMessageDisplay( "#xanMessage", "<span class='text-success'>" + params[ "Msg" ] + "<br />" + Math.abs( timeSecs ) + "s</span>", messageID, iconID, true, true );
                 },
                 error: function ( xhr, status, error ) {
-                    xanMessageDisplay( "#xanMessage", "<?= FA_PSOS_ERROR ?>", '', '', false, false );
+                    xanMessageDisplay( "#xanMessage", "<?= FI_PSOS_ERROR ?>", '', '', false, false );
                     xanMessageDisplay( "#xanMessage", "<span class='text-danger'>" + params[ "Msg" ] + " Error:" + error + "</span>", '', '', false, true );
                 }
             } );
@@ -576,7 +578,7 @@ ob_start();
 
         function xanDoSave( event ) {
             let iconID = Math.floor( ( Math.random() * 999999999 ) + 1 );
-            xanMessageDisplay( "#xanMessage", "<?= FA_PSOS_SAVE ?>", iconID, '', false, false );
+            xanMessageDisplay( "#xanMessage", "<?= FI_PSOS_SAVE ?>", iconID, '', false, false );
             thisInput = $( this );
             let thisInputID = $( thisInput ).attr( "id" );
             let thisInputName = $( thisInput ).attr( "name" );
@@ -639,7 +641,7 @@ ob_start();
                     },
                     error: function ( xhr, status, error ) {
                         //alert( "Error: " + status + " / " + error );
-                        xanMessageDisplay( "#xanMessage", "<?= FA_PSOS_ERROR ?>", '', '', false, false );
+                        xanMessageDisplay( "#xanMessage", "<?= FI_PSOS_ERROR ?>", '', '', false, false );
                         xanMessageDisplay( "#xanMessage", "<span class='text-danger'>Saving Error: " + error + "</span>", '', '', false, true );
                     }
                 } );
@@ -680,7 +682,7 @@ ob_start();
 
             // Message Display
             let iconID = Math.floor( ( Math.random() * 999999999 ) + 1 );
-            xanMessageDisplay( "#xanMessage", "<?= FA_PSOS_STOPWATCH ?>", iconID, '', true, false );
+            xanMessageDisplay( "#xanMessage", "<?= FI_PSOS_STOPWATCH ?>", iconID, '', true, false );
             let messageID = Math.floor( ( Math.random() * 999999999 ) + 1 );
             xanMessageDisplay( "#xanMessage", "<span class='text-success'>Load Page<br /><?= \xan\microsecsDiff( $pageload_begin ) ?></span>", messageID, iconID, true, true );